Move time conversion code out of __init__.py
